requirements.txt
telethon==1.36.0
aiosqlite==0.19.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
